requirements.txt
requests
python-dotenv
psycopg2-binary==2.9.7
fastapi==0.100.0
uvicorn==0.22.0
openai>=1.46.0

